Break function into two.
I broke the `parse_data`into two.

One build the time series of daily gain from the data obtained from Udacity.

The other one fills this data with zero values from one year ago up to
next Saturday. The objective of this one is to fill the time series with
371 values in order to be able to build the plot.